+++ src/Silverback.Integration.Kafka/Messaging/Messages/KafkaMessageHeaders.cs
@@ -12,12 +12,6 @@ namespace Silverback.Messaging.Messages;
 public static class KafkaMessageHeaders
 {
     /// <summary>
-    ///     The header that will be filled with the key of the message consumed from Kafka. It is also used to
-    ///     temporary store the key for the produced message, defined via <see cref="KafkaKeyMemberAttribute" />.
-    /// </summary>
-    public const string KafkaMessageKey = "x-kafka-message-key";
-
-    /// <summary>
     ///     The header that will be filled with the timestamp of the message consumed from Kafka.
     /// </summary>
     public const string Timestamp = "x-kafka-message-timestamp";